console/scale/shared: clean warnings as much as possible for now

+++ core/src/bambu/tray.rs
@@ -329,7 +329,7 @@ impl BambuPrinter {
     }
 
     pub fn set_tray_spool_rec(&mut self, tray_id: usize, spool_rec: &SpoolRecord) {
-        self.update_any_tray(tray_id as usize, |tray| {
+        self.update_any_tray(tray_id, |tray| {
             tray.meta_info = TrayMetaInfo::default();
             tray.meta_info.spool_id = Some(spool_rec.id.clone());
             tray.meta_info.consumed_since_weight = spool_rec.consumed_since_weight;
